# backend/docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ALPACART API",
    "description": "API Backend de Alpacart — Marca premium de alpaca peruana",
    "version": "1.0.0",
    "contact": {
      "name": "Alpacart Textiles",
      "url": "https://alpacart.com"
    }
  },
  "servers": [
    { "url": "http://localhost:8000/api/v1", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearer": [] }],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Iniciar sesión",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "mateo.q@alpacart.com" },
                  "password": { "type": "string", "minLength": 6, "example": "securePassword123" },
                  "remember": { "type": "boolean", "default": false }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Login exitoso, retorna accessToken y refreshToken" },
          "401": { "description": "Credenciales incorrectas" }
        }
      }
    },
...
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Perfil del usuario autenticado",
        "security": [{ "bearer": [] }],
        "responses": {
          "200": { "description": "Perfil del usuario" },
          "401": { "description": "No autorizado" }
        }
      }
    },
    "/products": {
      "get": {
        "tags": ["Catalog"],
        "summary": "Listar productos",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "perPage", "in": "query", "schema": { "type": "integer", "default": 25 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "categoryId", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Lista paginada de productos" } }
      },
      "post": {
        "tags": ["Catalog"],
        "summary": "Crear producto",
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "sku": { "type": "string" },
                  "price": { "type": "number" },
                  "categoryId": { "type": "string" }
                },
                "required": ["name", "sku", "price"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Producto creado" } }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": ["Catalog"],
        "summary": "Obtener producto",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Producto encontrado" }, "404": { "description": "No encontrado" } }
      },
      "put": {
        "tags": ["Catalog"],
        "summary": "Actualizar producto",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Producto actualizado" } }
      },
      "delete": {
        "tags": ["Catalog"],
        "summary": "Eliminar producto",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Producto eliminado" } }
      }
    },
...
    "/variants": {
      "post": {
        "tags": ["Catalog"],
        "summary": "Crear variante",
        "security": [{ "bearer": [] }],
        "responses": { "201": { "description": "Variante creada" } }
      }
    },
    "/variants/{id}": {
      "put": {
        "tags": ["Catalog"],
        "summary": "Actualizar variante",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Variante actualizada" } }
      },
      "delete": {
        "tags": ["Catalog"],
        "summary": "Eliminar variante",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Variante eliminada" } }
      }
    },
    "/products/{id}/media": {
      "post": {
        "tags": ["Catalog"],
        "summary": "Subir imagen",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "201": { "description": "Imagen subida" } }
      }
    },
...
    "/contents": {
      "get": {
        "tags": ["CMS"],
        "summary": "Contenido público",
        "responses": { "200": { "description": "Lista de contenidos" } }
      },
      "post": {
        "tags": ["CMS"],
        "summary": "Crear contenido",
        "security": [{ "bearer": [] }],
        "responses": { "201": { "description": "Contenido creado" } }
      }
    },
    "/contents/{id}": {
      "put": {
        "tags": ["CMS"],
        "summary": "Actualizar contenido",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Contenido actualizado" } }
      },
      "delete": {
        "tags": ["CMS"],
        "summary": "Eliminar contenido",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Contenido eliminado" } }
      }
    },
...
    "/admin/hero-slides": {
      "get": {
        "tags": ["CMS"],
        "summary": "Listar hero slides (admin)",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Lista de hero slides" } }
      },
      "post": {
        "tags": ["CMS"],
        "summary": "Crear hero slide",
        "security": [{ "bearer": [] }],
        "responses": { "201": { "description": "Hero slide creado" } }
      }
    },
    "/admin/hero-slides/{id}": {
      "put": {
        "tags": ["CMS"],
        "summary": "Actualizar hero slide",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Hero slide actualizado" } }
      },
      "delete": {
        "tags": ["CMS"],
        "summary": "Eliminar hero slide",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Hero slide eliminado" } }
      }
    },
    "/admin/gallery": {
      "get": {
        "tags": ["CMS"],
        "summary": "Listar galería (admin)",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Lista de galería" } }
      },
      "post": {
        "tags": ["CMS"],
        "summary": "Crear imagen galería",
        "security": [{ "bearer": [] }],
        "responses": { "201": { "description": "Imagen creada" } }
      }
    },
    "/admin/gallery/{id}": {
      "put": {
        "tags": ["CMS"],
        "summary": "Actualizar imagen",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Imagen actualizada" } }
      },
      "delete": {
        "tags": ["CMS"],
        "summary": "Eliminar imagen",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Imagen eliminada" } }
      }
    },
    "/admin/testimonials": {
      "get": {
        "tags": ["CMS"],
        "summary": "Listar testimonios (admin)",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Lista de testimonios" } }
      },
      "post": {
        "tags": ["CMS"],
        "summary": "Crear testimonio",
        "security": [{ "bearer": [] }],
        "responses": { "201": { "description": "Testimonio creado" } }
      }
    },
    "/admin/testimonials/{id}": {
      "put": {
        "tags": ["CMS"],
        "summary": "Actualizar testimonio",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Testimonio actualizado" } }
      },
      "delete": {
        "tags": ["CMS"],
        "summary": "Eliminar testimonio",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Testimonio eliminado" } }
      }
    },
    "/admin/benefits": {
      "get": {
        "tags": ["CMS"],
        "summary": "Listar beneficios (admin)",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Lista de beneficios" } }
      },
      "post": {
        "tags": ["CMS"],
        "summary": "Crear beneficio",
        "security": [{ "bearer": [] }],
        "responses": { "201": { "description": "Beneficio creado" } }
      }
    },
    "/admin/benefits/{id}": {
      "put": {
        "tags": ["CMS"],
        "summary": "Actualizar beneficio",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Beneficio actualizado" } }
      },
      "delete": {
        "tags": ["CMS"],
        "summary": "Eliminar beneficio",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Beneficio eliminado" } }
      }
    },
    "/admin/artisan-processes": {
      "get": {
        "tags": ["CMS"],
        "summary": "Listar procesos (admin)",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Lista de procesos" } }
      },
      "post": {
        "tags": ["CMS"],
        "summary": "Crear proceso",
        "security": [{ "bearer": [] }],
        "responses": { "201": { "description": "Proceso creado" } }
      }
    },
    "/admin/artisan-processes/{id}": {
      "put": {
        "tags": ["CMS"],
        "summary": "Actualizar proceso",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Proceso actualizado" } }
      },
      "delete": {
        "tags": ["CMS"],
        "summary": "Eliminar proceso",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Proceso eliminado" } }
      }
    }
  }
}